+++ @cybernaut/types/src/Condition.ts
@@ -1,9 +1,9 @@
 import {Accessor} from './Accessor';
 import {Predicate} from './Predicate';
 
-export interface Condition<T> {
+export interface Condition {
   readonly description: string;
-  readonly accessor: Accessor<T>;
-  readonly predicate: Predicate<T>;
+  readonly accessor: Accessor;
+  readonly predicate: Predicate;
   readonly negated: boolean;
 }